refactor(store): change the authentication flow
Wrap the logic of setting the auth data in local storage and move it
to the authSuccess effect. This way same auth logic can be executed for
log in as well as for other scenarios where saving the new values is needed such as
update user password where a new auth session and token are issued by the server.
Create more actions and effects to separate and handle different events