feat: implement update current user password functionality
Users can update their password in the user-dashboard.
Create new users module actions, effects, alter the users state and add cases to the reducer.
Create new method in users service. Use reactive forms. When user's password is updated a new jwt
is issued and the new auth data needs to be saved, therefore the authSuccess action is fired to re-run
the authentication logic again and keep the user signed in