feat: user image upload visual feedback
Displays the name of the file with a green check symbol if loaded
successfully. If file is not of type image displays an error notification
and disables the form submit until valid file is loaded. Works only if a file
is being selected, otherwise the user can update their info without problems